Update dependencies: Switch to `nrf-pac`, and update `bt-hci` and `embassy-time-driver` (#7)
* switch to nrf-pac

* update embassy-time, and embassy-time-driver

since the `allocate_alarm`, `set_alarm`, and `set_alarm_callback`
functions have been removed, the implementations for the
`ble_npl_callout_*` functions have been updated to use wakers.

unfortunately, this requires the `_generic_queue` feature to be enabled
on `embassy_time_queue_utils` in order for callouts to fully work. the
default `queue_integrated` implementation depends on `embassy_executor`,
which would panic since we are scheduling a waker that was not created
by `embassy_executor`. this may need to be revisited in the future.

on another note: the original implementations of the callout functions
were incorrect (the `event` field was a pointer type when it shouldn't
have been). this was also fixed as part of this commit.